+++ src/RemoteTech/RTCore.cs
@@ -18,6 +18,7 @@ namespace RemoteTech
 
         /// Addons
         public AddOns.ControlLockAddon ctrlLockAddon { get; protected set; }
+        public AddOns.KerbalAlarmClockAddon kacAddon { get; protected set; }
 
         public event Action OnFrameUpdate = delegate { };
         public event Action OnPhysicsUpdate = delegate { };
@@ -25,11 +26,13 @@ namespace RemoteTech
 
         public FilterOverlay FilterOverlay { get; protected set; }
         public FocusOverlay FocusOverlay { get; protected set; }
+        public ManeuverNodeOverlay ManeuverNodeOverlay { get; protected set; }
         public TimeWarpDecorator TimeWarpDecorator { get; protected set; }
 
         public void Start()
         {
-            if (Instance != null)
+            // Destroy the Core instance if != null or if Remotetech is disabled
+            if (Instance != null || !RTSettings.Instance.RemoteTechEnabled)
             {
                 Destroy(this);
                 return;
@@ -38,6 +41,7 @@ namespace RemoteTech
             Instance = this;
 
             ctrlLockAddon = new AddOns.ControlLockAddon();
+            kacAddon = new AddOns.KerbalAlarmClockAddon();
 
             Satellites = new SatelliteManager();
             Antennas = new AntennaManager();
@@ -46,6 +50,7 @@ namespace RemoteTech
 
             FilterOverlay = new FilterOverlay();
             FocusOverlay = new FocusOverlay();
+            ManeuverNodeOverlay = new ManeuverNodeOverlay();
             TimeWarpDecorator = new TimeWarpDecorator();
 
             FlightUIPatcher.Patch();
@@ -113,6 +118,10 @@ namespace RemoteTech
             if (Network != null) Network.Dispose();
             if (Satellites != null) Satellites.Dispose();
             if (Antennas != null) Antennas.Dispose();
+
+            // addons
+            if (ctrlLockAddon != null) ctrlLockAddon = null;
+            if (kacAddon != null) kacAddon = null;
 
             Instance = null;
         }
@@ -211,15 +220,36 @@ namespace RemoteTech
         public new void Start()
         {
             base.Start();
-            FilterOverlay.OnEnterMapView();
-            FocusOverlay.OnEnterMapView();
+            if(RTCore.Instance != null)
+            {
+                base.FilterOverlay.OnEnterMapView();
+                base.FocusOverlay.OnEnterMapView();
+                base.ManeuverNodeOverlay.OnEnterMapView();
+            }
         }
 
         private new void OnDestroy()
         {
-            FilterOverlay.OnExitMapView();
-            FocusOverlay.OnExitMapView();
+            if (RTCore.Instance != null)
+            {
+                base.FilterOverlay.OnExitMapView();
+                base.FocusOverlay.OnExitMapView();
+                base.ManeuverNodeOverlay.OnExitMapView();
+            }
             base.OnDestroy();
+        }
+    }
+
+    [KSPAddon(KSPAddon.Startup.MainMenu, false)]
+    public class RTMainMenu : MonoBehaviour
+    {
+        public void Start()
+        {
+            // Set the loaded trigger to false, this we will load a new
+            // settings after selecting a save game. This is necessary
+            // for switching between saves without shutting down the ksp
+            // instance.
+            RTSettings.Instance.settingsLoaded = false;
         }
     }
 }